# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ListSortOrder": {
        "description": "Sort direction accepted by every list endpoint.",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "ListSortOrder",
        "type": "string"
      },
      "PromptBlockModel": {
        "additionalProperties": false,
        "description": "Create/update payload for a reusable prompt block.",
        "properties": {
          "content": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_dynamic": {
            "default": false,
            "title": "Is Dynamic",
            "type": "boolean"
          },
          "is_public": {
            "default": false,
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "type": {
            "$ref": "#/components/schemas/PromptBlockType"
          }
        },
        "required": [
          "name",
          "content",
          "type"
        ],
        "title": "PromptBlockModel",
        "type": "object"
      },
      "PromptBlockPublic": {
        "description": "Public prompt block representation.",
        "properties": {
          "content": {
            "maxLength": 5000,
            "title": "Content",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_dynamic": {
            "default": false,
            "title": "Is Dynamic",
            "type": "boolean"
          },
          "is_public": {
            "default": false,
            "title": "Is Public",
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "slug": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Slug",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/PromptBlockType"
          }
        },
        "required": [
          "name",
          "slug",
          "content",
          "type",
          "id",
          "owner_id"
        ],
        "title": "PromptBlockPublic",
        "type": "object"
      },
      "PromptBlockSearchField": {
        "description": "Columns ``q`` may scan on ``GET /prompt-block/``.",
        "enum": [
          "name",
          "slug",
          "description",
          "content"
        ],
        "title": "PromptBlockSearchField",
        "type": "string"
      },
      "PromptBlockSortField": {
        "description": "Columns ``sort`` may order by on ``GET /prompt-block/``.",
        "enum": [
          "id",
          "name",
          "slug",
          "type",
          "is_dynamic",
          "is_public",
          "created_at",
          "updated_at"
        ],
        "title": "PromptBlockSortField",
        "type": "string"
      },
      "PromptBlockType": {
        "description": "Supported prompt block roles.",
        "enum": [
          "role",
          "task",
          "context",
          "instruction",
          "example",
          "format"
        ],
        "title": "PromptBlockType",
        "type": "string"
      },
      "PromptBlocksExport": {
        "description": "Unpaginated prompt block export, capped at ``MAX_EXPORT_SIZE`` (`A-C8`).\n\n``count`` is the filtered set's true size, same as ``PromptBlocksPublic``.\n``truncated`` is ``True`` only when the filtered set exceeds the cap and\n``data`` therefore holds fewer rows than ``count`` — the signal a caller\nneeds to narrow the filter rather than assume the export is complete.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/PromptBlockPublic"
            },
            "title": "Data",
            "type": "array"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "count",
          "truncated"
        ],
        "title": "PromptBlocksExport",
        "type": "object"
      },
      "PromptBlocksPublic": {
        "description": "Paginated prompt block list.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/PromptBlockPublic"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data",
          "count"
        ],
        "title": "PromptBlocksPublic",
        "type": "object"
      },
...
      "ResponseErrorBase": {
        "description": "Structured error response returned on exceptions.",
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ResponseError"
            },
            "title": "Errors",
            "type": "array"
          },
          "from_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Error"
          },
          "msg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msg"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          },
          "success": {
            "default": false,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ResponseErrorBase",
        "type": "object"
      },
      "ResponseMessage": {
        "description": "Simple success + message response.",
        "properties": {
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "msg"
        ],
        "title": "ResponseMessage",
        "type": "object"
      },
      "ResponseModelBase": {
        "description": "Generic success/data response wrapper.",
        "properties": {
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "ResponseModelBase",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/user/login/access-token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "prompt-engine-m8",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/prompt/prompt-block/": {
      "get": {
        "description": "Retrieve prompt blocks visible to the current user.\n\n``q`` searches every declared block column, or only ``csrc`` when one is\nnamed; ``f`` carries comma-joined facet values combined with ``OR``;\n``sort``/``order`` order the page. Every value is allow-listed by\n``ListQueryController`` — an undeclared one is a ``422``.\n\n``count`` is the count of the **filtered** set, not of everything visible.\nThat is the point of the parameters: a paginator driven by an unfiltered\ncount reports pages that do not exist. With none of the new parameters\npresent the response is byte-for-byte what ``skip``/``limit`` always\nreturned, ordering included — an absent ``sort`` adds no ``ORDER BY``.",
        "operationId": "prompt-block-prompt_block_list",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "csrc",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PromptBlockSearchField"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrc"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PromptBlockSortField"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ListSortOrder"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order"
            }
          },
          {
            "description": "Comma-joined facet values, combined with OR. Allowed: role, task, context, instruction, example, format, dynamic, static, public, private.",
            "in": "query",
            "name": "f",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-joined facet values, combined with OR. Allowed: role, task, context, instruction, example, format, dynamic, static, public, private.",
              "maxLength": 200,
              "title": "F",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptBlocksPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseErrorBase"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Prompt Block List",
        "tags": [
          "prompt-block"
        ]
      }
    },
    "/prompt/prompt-block/add/": {
      "post": {
        "description": "Create a prompt block.",
        "operationId": "prompt-block-add_prompt_block",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptBlockModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModelBase"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseErrorBase"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Add Prompt Block",
        "tags": [
          "prompt-block"
        ]
      }
    },
    "/prompt/prompt-block/delete/{item_id}/": {
      "delete": {
        "description": "Delete a prompt block.",
        "operationId": "prompt-block-delete_prompt_block",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseErrorBase"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Prompt Block",
        "tags": [
          "prompt-block"
        ]
      }
    },
    "/prompt/prompt-block/edit/{item_id}/": {
      "put": {
        "description": "Update a prompt block.",
        "operationId": "prompt-block-update_prompt_block",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptBlockModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModelBase"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseErrorBase"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Prompt Block",
        "tags": [
          "prompt-block"
        ]
      }
    },
    "/prompt/prompt-block/export/": {
      "get": {
        "description": "Export every prompt block in the filtered set, not one page of it (`A-C8`).\n\nCarries the same ``q``/``csrc``/``sort``/``order``/``f`` vocabulary as\n``GET /prompt-block/`` — deliberately no ``skip``/``limit``, since paging\na bulk export defeats the point of it. Bounded instead by\n``MAX_EXPORT_SIZE``: a filtered set larger than that returns the first\n``MAX_EXPORT_SIZE`` rows with ``truncated: true`` rather than materialising\nan unbounded result.",
        "operationId": "prompt-block-prompt_block_export",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "csrc",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PromptBlockSearchField"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrc"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PromptBlockSortField"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ListSortOrder"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Order"
            }
          },
          {
            "description": "Comma-joined facet values, combined with OR. Allowed: role, task, context, instruction, example, format, dynamic, static, public, private.",
            "in": "query",
            "name": "f",
            "required": false,
            "schema": {
              "default": "",
              "description": "Comma-joined facet values, combined with OR. Allowed: role, task, context, instruction, example, format, dynamic, static, public, private.",
              "maxLength": 200,
              "title": "F",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptBlocksExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseErrorBase"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Prompt Block Export",
        "tags": [
          "prompt-block"
        ]
      }
    },
    "/prompt/prompt-block/get/{item_id}/": {
      "get": {
        "description": "Get a prompt block by ID.",
        "operationId": "prompt-block-get_prompt_block",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModelBase"
                    },
                    {
                      "$ref": "#/components/schemas/ResponseMessage"
                    }
                  ],
                  "title": "Response Prompt-Block-Get Prompt Block"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseErrorBase"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Prompt Block",
        "tags": [
          "prompt-block"
        ]
      }
    },
    "/prompt/prompt-block/get_by_slug/{item_slug}/": {
      "get": {
        "description": "Get a prompt block by slug.",
        "operationId": "prompt-block-get_prompt_block_by_slug",
        "parameters": [
          {
            "in": "path",
            "name": "item_slug",
            "required": true,
            "schema": {
              "title": "Item Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ResponseModelBase"
                    },
                    {
                      "$ref": "#/components/schemas/ResponseMessage"
                    }
                  ],
                  "title": "Response Prompt-Block-Get Prompt Block By Slug"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseErrorBase"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Prompt Block By Slug",
        "tags": [
          "prompt-block"
        ]
      }
    },
...
  }
}